Grabbed data from matchData.csv
Over 40,017 matches grabbed this script grabs all the matches and stores the champion, runes used, kda and other relevant statistics. This process took over 24 hours as I was limited to 100 requests per 2 minutes. However, the data collected still needs some cleaning up as some information gathered was before the new rune system was implemented so perks will be : "NA". This data needs to be removed. Also, I did not check for duplicate matches and thus there may be duplicate rows in the matchStats. I assume that if two rows contain the same values they are duplicates as it would be highly unlikely that 2 players have the exact same values because many are continuous.